Dépot temporaire, positionnement des robots et paticules respectés mais la carte n'est pas entièrement affichée

+++ rendu 1/Robot.cc
@@ -104,7 +104,7 @@ int Robot_N::get_k_update_panne()
     return m_k_update_panne;
 }
 
-int Robot_N::get_angle()
+double Robot_N::get_angle()
 {
     return m_angle;
 }
